Bitsy Board Petting Zoo
The petting zoo is a project where I created 3 distinct "animals" on a small breadboard using an Attiiny85.
This repository contains the Arduino code, circuit schematic, PCB design, BOM for the PCBs, and a video of each "animal."
OwlTiny85: The OwlTiny85 acts like most owls, he is nocturnal. As it gets darker in the room, he awakens progressivley.
Whipporwhill: The Whipporwhill is somewhat like his real life counter part in that he will sing you a song. However, he is somewhat shy in his bird house, so you have to knock for him to sing to you.
Mecha Bird Drone:  As we all knowbirds are just government spies. This one is no exception, and the potentiometer that protrudes from his side is proof. By turning this dial you can tune into any "frequency" to spy on others.